Recloned files and added in test case.
Had to reclone the files and build the project due to a bug in Intellij, the IDE I am using. It refused to build the latest versions of the files. Added in a test case for creating the DAG and adding edges. Function defintions were incorrect so fixed those in LowestCommonAncestorTest